// src/enums/env.enum.ts
export enum EEnv {
  PORT = 'PORT',

  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_DATABASE = 'DB_DATABASE',

  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',

  RPC_URL = 'RPC_URL',
  MAX_BLOCK_RANGE = 'MAX_BLOCK_RANGE',
  MIN_BLOCK_RANGE = 'MIN_BLOCK_RANGE',
  SAFE_BLOCK = 'SAFE_BLOCK',
  START_BLOCK = 'START_BLOCK',
}